(PC-17714)[API] refactor: move provider commands
